Fix code review recommendations
