+++ tests/system/action/speaker/test_create.py
@@ -21,7 +21,7 @@ class SpeakerCreateActionTest(BaseActionTestCase):
         speaker = self.get_model("speaker/1")
         assert speaker.get("user_id") == 7
         assert speaker.get("list_of_speakers_id") == 23
-        assert speaker.get("weight") == 10000
+        assert speaker.get("weight") == 1
         list_of_speakers = self.get_model("list_of_speakers/23")
         assert list_of_speakers.get("speaker_ids") == [1]
         user = self.get_model("user/7")
@@ -215,3 +215,89 @@ class SpeakerCreateActionTest(BaseActionTestCase):
         self.assertIn(
             "Only present users can be on the lists of speakers.", str(response.data)
         )
+
+    def test_create_standard_speaker_in_only_talker_list(self) -> None:
+        self.create_model("meeting/7844", {"name": "name_asdewqasd"})
+        self.create_model("user/7", {"username": "talking"})
+        self.create_model(
+            "speaker/1",
+            {
+                "user_id": 7,
+                "list_of_speakers_id": 23,
+                "begin_time": 100000,
+                "weight": 5,
+            },
+        )
+        self.create_model(
+            "list_of_speakers/23", {"speaker_ids": [1], "meeting_id": 7844}
+        )
+        response = self.client.post(
+            "/",
+            json=[
+                {
+                    "action": "speaker.create",
+                    "data": [{"user_id": 1, "list_of_speakers_id": 23}],
+                }
+            ],
+        )
+        self.assert_status_code(response, 200)
+        self.assert_model_exists(
+            "speaker/2",
+            {"user_id": 1, "weight": 1},
+        )
+        self.assert_model_exists("list_of_speakers/23", {"speaker_ids": [1, 2]})
+
+    def test_create_standard_speaker_at_the_end_of_filled_list(self) -> None:
+        self.create_model("meeting/7844", {"name": "name_asdewqasd"})
+        self.create_model("user/7", {"username": "talking"})
+        self.create_model("user/8", {"username": "waiting"})
+        self.update_model(
+            "user/1",
+            {
+                "speaker_$7844_ids": [3],
+                "speaker_$_ids": ["7844"],
+            },
+        )
+        self.create_model(
+            "speaker/1",
+            {
+                "user_id": 7,
+                "list_of_speakers_id": 23,
+                "begin_time": 100000,
+                "weight": 5,
+            },
+        )
+        self.create_model(
+            "speaker/2", {"user_id": 8, "list_of_speakers_id": 23, "weight": 1}
+        )
+        self.create_model(
+            "speaker/3",
+            {
+                "user_id": 1,
+                "list_of_speakers_id": 23,
+                "point_of_order": True,
+                "weight": 2,
+            },
+        )
+        self.create_model(
+            "list_of_speakers/23", {"speaker_ids": [1, 2, 3], "meeting_id": 7844}
+        )
+        response = self.client.post(
+            "/",
+            json=[
+                {
+                    "action": "speaker.create",
+                    "data": [{"user_id": 1, "list_of_speakers_id": 23}],
+                }
+            ],
+        )
+        self.assert_status_code(response, 200)
+        self.assert_model_exists(
+            "speaker/3",
+            {"user_id": 1, "point_of_order": True, "weight": 2},
+        )
+        self.assert_model_exists(
+            "speaker/4",
+            {"user_id": 1, "point_of_order": None, "weight": 3},
+        )
+        self.assert_model_exists("list_of_speakers/23", {"speaker_ids": [1, 2, 3, 4]})
